+++ src/cpp/server/ServerLogVars.cpp
@@ -17,7 +17,7 @@
 
 #include <core/LogOptions.hpp>
 
-#include <server/ServerSessionManager.hpp>
+#include <server/session/ServerSessionManager.hpp>
 
 using namespace rstudio::core;
