+++ src/cpp/server/ServerPAMAuth.cpp
@@ -28,10 +28,11 @@
 
 #include <server/ServerOptions.hpp>
 #include <server/ServerUriHandlers.hpp>
-#include <server/ServerSessionProxy.hpp>
 
 #include <server/auth/ServerAuthHandler.hpp>
 #include <server/auth/ServerAuthCommon.hpp>
+
+#include <server/session/ServerSessionProxy.hpp>
 
 #include "ServerLoginPages.hpp"
